refactor c++ code
