+++ src/main/cc/simif_emul.h
@@ -4,7 +4,7 @@
 #include "simif.h"
 #include "mm.h"
 #include "mm_dramsim2.h"
-#include "mmio.h"
+#include "emul/mmio.h"
 
 class simif_emul_t : public virtual simif_t
 {
